feat: permissions data source
This adds a `snyk_permissions` resource.  I first tried doing it as a data
resource, but that doesn't quite work as the template we're returning is
non-deterministic in the role name.

This makes it a lot easier to set up environments, since you don't need the
manual steps of generating a template and getting the role ARN.

Example usage:

    resource "snyk_permissions" "aws_cf_permissions" {
      organization_id = var.snyk_org
      platform        = "aws"
      type            = "cf"
    }

    resource "aws_cloudformation_stack" "rolestack" {
      name          = "tf-snyk-dev-role"
      capabilities  = ["CAPABILITY_NAMED_IAM"]
      template_body = snyk_permissions.aws_cf_permissions.data
    }

    resource "snyk_environment" "example" {
      name            = "little_env_12345_lol"
      kind            = "aws"
      organization_id = var.snyk_org
      aws {
        role_arn = aws_cloudformation_stack.rolestack.outputs.SnykCloudRoleArn
      }
    }